+++ golden-paths/h1-foundation/web-application/skeleton/backend/aspnet/obj/Debug/net8.0/app.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("app")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+f5fd0945de11769f7a8f9a75f8be636422c9766f")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+5828636cf7e4a9ec2b9e4480bc534078294e162c")]
 [assembly: System.Reflection.AssemblyProductAttribute("app")]
 [assembly: System.Reflection.AssemblyTitleAttribute("app")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]